refactor: Order 서비스에서 Menu 의존성 제거
